Check valid dest on producer init when specified
Previoulsy when a client created a JMS MessageProducer using the
OpenWire protocol, the server would check to ensure that the destination
existed.  If it did not it would throw an invalid destination exception.
This did not take into account that the createProducer method on the
Session could be passed null when a destination is not specified.  This
patch updates the AMQProducer init method to ensure that the destination
is only validated when one is specified i.e. when null is specified no
check takes place.